refactor: split server into route modules with decorator-based auth
Replace the monolithic litestar_app.py with:
- server/auth.py: AuthMiddleware + @read_only decorator (no hardcoded path list)
- server/routes/search.py: search, ask, ask_stream, chat, chat_stream
- server/routes/models.py: catalog, installed, pull, show, delete, set_chat/vision
- server/routes/documents.py: add, list, remove, sync
- server/routes/crawl.py: crawl stream endpoint
- server/app.py: app factory importing routes from modules

litestar_app.py kept as thin re-export shim for backwards compatibility.
