+++ src/lilbee/server/routes/search.py
@@ -1,0 +1,96 @@
+"""Search, ask, ask_stream, chat, and chat_stream route handlers."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from litestar import get, post
+from litestar.params import Parameter
+from litestar.response import Stream
+
+from lilbee.query import ChatMessage as ChatMessageDict
+from lilbee.server import handlers
+from lilbee.server.auth import read_only
+from lilbee.server.models import (
+    AskRequest,
+    AskResponse,
+    ChatRequest,
+    CleanedChunk,
+)
+
+
+def _clean_to_model(raw: dict) -> CleanedChunk:
+    """Convert a raw cleaned dict to a CleanedChunk model."""
+    return CleanedChunk(**raw)
+
+
+@get("/api/search")
+@read_only
+async def search_route(
+    q: str = Parameter(query="q"),
+    top_k: int = Parameter(query="top_k", default=5),
+) -> list[dict[str, Any]]:
+    """Search indexed documents by semantic similarity. No LLM call required."""
+    return await handlers.search(q, top_k=top_k)
+
+
+@post("/api/ask")
+async def ask_route(data: AskRequest) -> AskResponse:
+    """One-shot RAG question returning an answer with source chunks."""
+    raw = await handlers.ask(
+        question=data.question,
+        top_k=data.top_k,
+        options=data.options,
+    )
+    return AskResponse(
+        answer=raw["answer"],
+        sources=[_clean_to_model(s) for s in raw["sources"]],
+    )
+
+
+@post("/api/ask/stream")
+async def ask_stream_route(data: AskRequest) -> Stream:
+    """Streaming SSE version of ask, emitting token-by-token answer chunks."""
+    return Stream(
+        handlers.ask_stream(
+            question=data.question,
+            top_k=data.top_k,
+            options=data.options,
+        ),
+        media_type="text/event-stream",
+    )
+
+
+@post("/api/chat")
+async def chat_route(data: ChatRequest) -> AskResponse:
+    """RAG chat with conversation history, returning an answer with sources."""
+    history: list[ChatMessageDict] = [
+        ChatMessageDict(role=m.role, content=m.content) for m in data.history
+    ]
+    raw = await handlers.chat(
+        question=data.question,
+        history=history,
+        top_k=data.top_k,
+        options=data.options,
+    )
+    return AskResponse(
+        answer=raw["answer"],
+        sources=[_clean_to_model(s) for s in raw["sources"]],
+    )
+
+
+@post("/api/chat/stream")
+async def chat_stream_route(data: ChatRequest) -> Stream:
+    """Streaming SSE version of chat with conversation history."""
+    history: list[ChatMessageDict] = [
+        ChatMessageDict(role=m.role, content=m.content) for m in data.history
+    ]
+    return Stream(
+        handlers.chat_stream(
+            question=data.question,
+            history=history,
+            top_k=data.top_k,
+            options=data.options,
+        ),
+        media_type="text/event-stream",
+    )
