+++ tests/server/test_handlers.py
@@ -11,13 +11,13 @@ import pytest
 from litestar.testing import AsyncTestClient
 
 from lilbee.config import cfg
-from lilbee.server import litestar_app as _litestar_mod
+from lilbee.server import auth as _auth_mod
 from lilbee.server.handlers import MAX_ADD_FILES
 
 
 def _auth_headers() -> dict[str, str]:
     """Return Authorization header using the current session token."""
-    return {"Authorization": f"Bearer {_litestar_mod._session_token}"}
+    return {"Authorization": f"Bearer {_auth_mod._session_token}"}
 
 
 @pytest.fixture(autouse=True)
@@ -78,7 +78,7 @@ class TestAddEndpoint:
         self, mock_extract_file, mock_embed_batch, mock_validate_model, isolated_env, tmp_path
     ):
         """POST /api/add with a valid file streams SSE events and adds it."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         src = tmp_path / "input.txt"
         src.write_text("Hello world content for testing.")
@@ -100,7 +100,7 @@ class TestAddEndpoint:
         self, mock_extract_file, mock_embed_batch, mock_validate_model, isolated_env, tmp_path
     ):
         """Nonexistent paths appear in the summary errors list."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         async with AsyncTestClient(create_app()) as client:
             resp = await client.post(
@@ -116,7 +116,7 @@ class TestAddEndpoint:
         self, mock_extract_file, mock_embed_batch, mock_validate_model, isolated_env, tmp_path
     ):
         """The force flag allows overwriting existing files."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         src = tmp_path / "dup.txt"
         src.write_text("Version 1")
@@ -136,7 +136,7 @@ class TestAddEndpoint:
         self, mock_extract_file, mock_embed_batch, mock_validate_model, isolated_env, tmp_path
     ):
         """The done event includes added, updated, removed, failed counts."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         src = tmp_path / "doc.txt"
         src.write_text("Content for done event testing.")
@@ -157,7 +157,7 @@ class TestAddEndpoint:
         self, mock_extract_file, mock_embed_batch, mock_validate_model, isolated_env, tmp_path
     ):
         """file_start event includes total_files and current_file."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         src = tmp_path / "progress.txt"
         src.write_text("Progress tracking test.")
@@ -176,7 +176,7 @@ class TestAddEndpoint:
         self, mock_extract_file, mock_embed_batch, mock_validate_model, isolated_env, tmp_path
     ):
         """Vision model parameter is temporarily set on cfg during sync."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         src = tmp_path / "doc.txt"
         src.write_text("Content for vision model test.")
@@ -197,7 +197,7 @@ class TestAddEndpoint:
 class TestAddValidation:
     async def test_empty_paths_returns_400(self, isolated_env):
         """POST /api/add with empty paths list returns 400."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         async with AsyncTestClient(create_app()) as client:
             resp = await client.post("/api/add", json={"paths": []}, headers=_auth_headers())
@@ -205,7 +205,7 @@ class TestAddValidation:
 
     async def test_missing_paths_returns_400(self, isolated_env):
         """POST /api/add without paths key returns 400."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         async with AsyncTestClient(create_app()) as client:
             resp = await client.post("/api/add", json={"force": True}, headers=_auth_headers())
@@ -213,7 +213,7 @@ class TestAddValidation:
 
     async def test_too_many_files_returns_400(self, isolated_env):
         """POST /api/add with >100 paths returns 400."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         paths = [f"/fake/file_{i}.txt" for i in range(MAX_ADD_FILES + 1)]
         async with AsyncTestClient(create_app()) as client:
@@ -222,7 +222,7 @@ class TestAddValidation:
 
     async def test_exactly_max_files_accepted(self, isolated_env, tmp_path):
         """POST /api/add with exactly 100 paths is accepted (paths can be nonexistent)."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         paths = [f"/fake/file_{i}.txt" for i in range(MAX_ADD_FILES)]
         with (
@@ -294,7 +294,7 @@ class TestOptionsPassthrough:
     """Verify generation options are extracted from request body and passed through."""
 
     async def test_ask_passes_options(self, isolated_env):
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         with (
             mock.patch("lilbee.embedder.embed", return_value=[0.1] * 768),
@@ -311,7 +311,7 @@ class TestOptionsPassthrough:
             assert "answer" in body
 
     async def test_chat_passes_options(self, isolated_env):
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         with (
             mock.patch("lilbee.embedder.embed", return_value=[0.1] * 768),
@@ -333,7 +333,7 @@ class TestOptionsPassthrough:
 
     async def test_ask_without_options(self, isolated_env):
         """Request without options field still works."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         with (
             mock.patch("lilbee.embedder.embed", return_value=[0.1] * 768),
@@ -349,7 +349,7 @@ class TestOptionsPassthrough:
 class TestCreateApp:
     def test_app_has_add_route(self):
         """The Litestar app registers the /api/add route."""
-        from lilbee.server.litestar_app import create_app
+        from lilbee.server.app import create_app
 
         app = create_app()
         paths = [r.path for r in app.routes]
